Consolida la corrección del tipo de dato de las llaves primarias en todo el sistema, migrando de int a Long y asegurando el uso de autoincremento mediante JPA. Los cambios se aplican de forma consistente en todas las capas:
entidades, repositorios, servicios y controladores.

Cambios incluidos:
Actualización de llaves primarias de int a Long y adición de autoincremento en las entidades Cliente, Usuario, Venta, DetalleVenta y Producto.
Corrección del tipo de dato de los identificadores en los repositorios correspondientes.
Ajuste de las interfaces y métodos de los servicios para trabajar con Long.
Actualización de los parámetros de las peticiones HTTP y PathVariables en los controladores para mantener coherencia de tipos.

+++ src/main/java/com/joseescobar/kinalapp/controller/DetalleVentaController.java
@@ -24,7 +24,7 @@ public class DetalleVentaController {
     }
 
     @GetMapping("/{codigoDetalleVenta}")
-    public ResponseEntity<DetalleVenta> buscarPorCodigo(@PathVariable int codigoDetalleVenta){
+    public ResponseEntity<DetalleVenta> buscarPorCodigo(@PathVariable Long codigoDetalleVenta){
         return detalleVentaService.buscarPorCodigo(codigoDetalleVenta).map(ResponseEntity::ok).orElse(ResponseEntity.notFound().build());
     }
 
@@ -39,7 +39,7 @@ public class DetalleVentaController {
     }
 
     @DeleteMapping("/{codigoDetalleVenta}")
-    public ResponseEntity<Void> eliminar(@PathVariable int codigoDetalleVenta){
+    public ResponseEntity<Void> eliminar(@PathVariable Long codigoDetalleVenta){
         try{
             if (!detalleVentaService.existePorCodigo(codigoDetalleVenta)){
                 return ResponseEntity.notFound().build();
@@ -52,7 +52,7 @@ public class DetalleVentaController {
     }
 
     @PutMapping("/{codigoDetalleVenta}")
-    public ResponseEntity<?> actualizar(@PathVariable int codigoDetalleVenta, @RequestBody DetalleVenta detalleVenta) {
+    public ResponseEntity<?> actualizar(@PathVariable Long codigoDetalleVenta, @RequestBody DetalleVenta detalleVenta) {
         try {
             DetalleVenta detalleVentaActualizado = detalleVentaService.actualizar(codigoDetalleVenta, detalleVenta);
             return ResponseEntity.ok(detalleVentaActualizado);
