+++ src/main/java/com/joseescobar/kinalapp/repository/DetalleVentaRepository.java
@@ -3,5 +3,5 @@ package com.joseescobar.kinalapp.repository;
 import com.joseescobar.kinalapp.entity.DetalleVenta;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DetalleVentaRepository extends JpaRepository<DetalleVenta, Integer> {
+public interface DetalleVentaRepository extends JpaRepository<DetalleVenta, Long> {
 }
